data-generation/requirements.txt
pandas
numpy
faker
pyyaml
pyarrow
sqlalchemy
pyodbc
python-dotenv